File reorganization and code cleanup
