+++ megamek/src/megamek/client/ui/swing/tileset/EntityImage.java
@@ -25,7 +25,7 @@ import java.util.Iterator;
 
 import megamek.MegaMek;
 import megamek.client.ui.swing.GUIPreferences;
-import megamek.client.ui.swing.util.ImageFileFactory;
+import megamek.common.util.fileUtils.ImageFileFactory;
 import megamek.common.*;
 import megamek.common.util.fileUtils.DirectoryItems;
 import megamek.common.util.ImageUtil;
