+++ megamek/src/megamek/common/util/fileUtils/AbstractDirectory.java
@@ -18,8 +18,6 @@
  */
 package megamek.common.util.fileUtils;
 
-import megamek.common.util.ItemFile;
-import megamek.common.util.ItemFileFactory;
 import megamek.common.util.StringUtil;
 
 import java.io.File;
